+++ xquic/src/main/cpp/main/include/xquic_h3_ctrl.h
@@ -17,7 +17,7 @@ ssize_t client_send_h3_content(xqc_cli_user_stream_t *user_stream);
 int client_send_h3_requests(xqc_cli_user_conn_t *user_conn,
                             xqc_cli_user_stream_t *user_stream, xqc_cli_request_t *req);
 
-int client_send_H3_ping(xqc_cli_user_conn_t *user_conn, char *ping_user_data);
+ssize_t client_send_H3_ping(xqc_cli_user_conn_t *user_conn, char *ping_user_data);
 
 #ifdef __cplusplus
 }
